Feature category products relation

+++ database/migrations/2019_11_18_152016_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('description');
             $table->decimal('unit_price');
             $table->decimal('stock');
-            $table->unsignedInteger('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->timestamps();
         });
     }
